refactor(components): update import paths and type definitions
- Changed UserProfile interface to a type alias for better clarity.
- Updated import paths for utility functions in ActivityList, TokensList, and CollectiblesList components.
